Redoing database helper methods
