+++ app/src/main/java/com/example/dealseeker_login/model/Product.kt
